# ДЗ_5/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Email System (User, Folder, Email)",
    "description": "Сводная OpenAPI документация по трём сервисам: User Service, Folder Service, Email Service.",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "auth",
      "description": "Авторизация (JWT)"
    },
    {
      "name": "users",
      "description": "Управление пользователями"
    },
    {
      "name": "folders",
      "description": "Управление почтовыми папками"
    },
    {
      "name": "emails",
      "description": "Управление письмами"
    }
  ],
  "paths": {
    "/auth/token": {
      "post": {
        "tags": ["auth"],
        "summary": "Получить JWT-токен по логину/паролю",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно получен токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Неверный логин/пароль"
          }
        }
      }
    },

    "/users": {
      "post": {
        "tags": ["users"],
        "summary": "Создать нового пользователя",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пользователь успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "400": {
            "description": "Логин уже занят / валидационная ошибка"
          }
        }
      },
      "get": {
        "tags": ["users"],
        "summary": "Получить список пользователей (или поиск)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Список пользователей",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  }
                }
              }
            }
          }
        }
      }
    },

    "/users/{user_id}": {
      "get": {
        "tags": ["users"],
        "summary": "Получить пользователя по ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о пользователе",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          }
        }
      }
    },

    "/folders": {
      "post": {
        "tags": ["folders"],
        "summary": "Создать новую папку",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Папка успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderRead"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["folders"],
        "summary": "Получить список всех папок",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Список папок",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FolderRead"
                  }
                }
              }
            }
          }
        }
      }
    },

    "/folders/{folder_id}": {
      "get": {
        "tags": ["folders"],
        "summary": "Получить папку по ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о папке",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderRead"
                }
              }
            }
          },
          "404": {
            "description": "Папка не найдена"
          }
        }
      },
      "patch": {
        "tags": ["folders"],
        "summary": "Обновить (переименовать) папку",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённая информация о папке",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderRead"
                }
              }
            }
          },
          "404": {
            "description": "Папка не найдена"
          }
        }
      },
      "delete": {
        "tags": ["folders"],
        "summary": "Удалить папку",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Папка удалена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Папка не найдена"
          }
        }
      }
    },

    "/emails": {
      "post": {
        "tags": ["emails"],
        "summary": "Создать новое письмо",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Письмо успешно создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailRead"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["emails"],
        "summary": "Получить список всех писем",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Список писем",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmailRead"
                  }
                }
              }
            }
          }
        }
      }
    },

    "/emails/{message_id}": {
      "get": {
        "tags": ["emails"],
        "summary": "Получить письмо по ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о письме",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailRead"
                }
              }
            }
          },
          "404": {
            "description": "Письмо не найдено"
          }
        }
      },
      "patch": {
        "tags": ["emails"],
        "summary": "Обновить письмо (например, пометить как прочитанное)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённое письмо",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailRead"
                }
              }
            }
          },
          "404": {
            "description": "Письмо не найдено"
          }
        }
      },
      "delete": {
        "tags": ["emails"],
        "summary": "Удалить письмо",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Письмо удалено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Письмо не найдено"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "TokenRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["username", "password"]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "token_type": { "type": "string", "default": "bearer" }
        },
        "required": ["access_token"]
      },

      "UserCreate": {
        "type": "object",
        "properties": {
          "login": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["login", "password"]
      },
      "UserRead": {
        "type": "object",
        "properties": {
          "user_id": { "type": "integer" },
          "login": { "type": "string" }
        }
      },

      "FolderCreate": {
        "type": "object",
        "properties": {
          "user_id": { "type": "integer" },
          "name": { "type": "string" }
        },
        "required": ["user_id", "name"]
      },
      "FolderRead": {
        "type": "object",
        "properties": {
          "folder_id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "name": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "FolderUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" }
        },
        "required": ["name"]
      },

      "EmailCreate": {
        "type": "object",
        "properties": {
          "from_user_id": { "type": "integer" },
          "to_user_id": { "type": "integer" },
          "subject": { "type": "string" },
          "body": { "type": "string" },
          "attachments": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["from_user_id", "to_user_id", "subject", "body"]
      },
      "EmailRead": {
        "type": "object",
        "properties": {
          "message_id": { "type": "integer" },
          "from_user_id": { "type": "integer" },
          "to_user_id": { "type": "integer" },
          "subject": { "type": "string" },
          "body": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "is_read": { "type": "boolean" },
          "attachments": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "EmailUpdate": {
        "type": "object",
        "properties": {
          "is_read": { "type": "boolean" }
        },
        "required": ["is_read"]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}